requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.5.0
pymongo>=4.8.0
pandas>=2.2.0
numpy>=2.0.0
prophet>=1.1.5
xgboost>=2.0.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
plotly>=5.22.0
python-multipart>=0.0.9
streamlit>=1.35.0
requests>=2.31.0
